[feature/only-op-can-view] Single permissions hiding
This commit applyies all the (somewhat) generic code needed to
create a system that allows permissions to be shown or hidden
based on the chosen option in other settings.
This is supposed to be used to hide an option where
choosing yes or choosing No or Never makes no difference based
on the Yes/No/Never of another option.
This includes a fix to IE8 that, according to IE rules, works
always.
Tested on IE7, IE8, IE9, FF 3.6, FF 10, Chrome (latest),
Opera (latest).

PHPBB3-10624