[feature/only-op-can-view] Performance improvement: Test the topic author first
This change is a simple change places. Initially the code was checking for the
hardest thing to do (between the two things that need to be checked), which is
to check if the user has f_read_other permissions. This allows the check if
the user is the author of the post before the check if he has the permission

This is just a performance improvement

PHPBB3-10623